requirements.txt
streamlit==1.26.0
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0